иправлен баг с динамической подгрузкой, логика поиска перенесена в новый компонент, StarIcon корректно окрашивается при добавлении и удалении из избранного, есть баг с тем что страница favotites не rerenderится после удаления фильма
