+++ src/supabase/UpdateFavorites.js
@@ -2,8 +2,6 @@ import supabase from './supabaseClient'
 
 
 const UpdateFavorites = async ( movieId, supaId, favoriteMovies, setFavoriteMovies) => {
-    console.log('movieId', movieId)
-    console.log('supaId', supaId)
     const { data: { user } } = await supabase.auth.getUser()
 
     const isFavorite = favoriteMovies.some((favorite) => favorite.movie === movieId)
@@ -15,10 +13,9 @@ const UpdateFavorites = async ( movieId, supaId, favoriteMovies, setFavoriteMovi
             .eq('id', supaId)
 
             console.log('delete')
-        if (data) {
-            const updatedFavorites = favoriteMovies.filter((favorite) => favorite.movie !== movieId)
+            const updatedFavorites = favoriteMovies.filter((favorite) => favorite.id !== supaId)
             setFavoriteMovies(updatedFavorites)
-        }
+        
     } else {
         const { data, error } = await supabase
             .from('favorites')
